+++ changelog.md
@@ -2,6 +2,15 @@
 
 Parsing and sanitization improvements are always ongoing and non-breaking
 
+## 1.7.x
+
+- Added `AirSigmet` report class and managers
+- Added `data` module to update station info client side
+- Added `flight_path` module for resolving 
+- Added `coord` to `Station`
+- Split `AVWXBase` some functionality into `ManagedReport`
+- Dropped 3.7 support
+
 ## 1.6.x
 
 - Added station text-based search at `station.search(text)`